\name{ant111b}
\alias{ant111b}
\title{Averages by block of corn yields, for treatment 111 only}
\description{
  These data frames have averages by blocks (parcels) for the treatment
  \code{111}.
}
\usage{ant111b}
\format{ A data frame with 36 observations on 9 variables.
  \describe{
    \item{site}{a factor with levels (\code{ant111b}:) \code{DBAN}
      \code{LFAN} \code{NSAN} \code{ORAN} \code{OVAN} \code{TEAN}
      \code{WEAN} \code{WLAN}
    }
    \item{parcel}{a factor with levels \code{I} \code{II} \code{III} \code{IV}}
    \item{code}{a numeric vector}
    \item{island}{a numeric vector}
    \item{id}{a numeric vector}
    \item{plot}{a numeric vector}
    \item{trt}{a numeric vector}
    \item{ears}{a numeric vector}
    \item{harvwt}{a numeric vector}
  }
}
\source{
  Andrews DF; Herzberg AM, 1985. Data. A Collection of Problems from
  Many Fields for the Student and Research Worker. Springer-Verlag.
  (pp. 339-353)
}
\keyword{datasets}
