\name{biomass}
\alias{biomass}
\non_function{}
\title{Biomass Data}
\usage{biomass}
\description{
The \code{biomass} data frame has 135 rows and 8 columns.  The
\code{rainforest} data frame is a subset of this one.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{dbh}{a numeric vector}
    \item{wood}{a numeric vector}
    \item{bark}{a numeric vector}
    \item{fac26}{a factor with 3 levels}
    \item{root}{a numeric vector}
    \item{rootsk}{a numeric vector}
    \item{branch}{a numeric vector}
    \item{species}{a factor with levels
\code{Acacia mabellae},
\code{C. fraseri},
\code{Acmena smithii},
\code{B. myrtifolia} }
  }
}
\source{
J. Ash, Australian National University
}
\references{
Ash, J. and Helman, C. (1990) Floristics and vegetation
biomass of a forest catchment, Kioloa, south coastal N.S.W.
Cunninghamia, 2: 167-182.
}
\keyword{datasets}
