\name{bostonc}
\alias{bostonc}
\docType{data}
\title{Boston Housing Data -- Corrected}
\usage{bostonc}
\description{
The corrected Boston housing data (from http://lib.stat.cmu.edu/datasets/).
}
\format{
  A single vector containing the contents of "boston_corrected.txt". 
}
\source{
Harrison, D. and Rubinfeld, D.L. 'Hedonic prices and the demand for clean air', J. Environ. Economics & Management, vol.5, 81-102, 1978.
corrected by Kelley Pace (kpace@unix1.sncc.lsu.edu)
}
\keyword{datasets}
