\name{DAAG-package}
\alias{DAAG-package}
\alias{DAAG}
\docType{package}
\title{
The R DAAG Package
}
\description{
Various data sets and functions used or referred to in the book
        Maindonald, J.H. and Braun, W.J. (3rd edn 2010)
        "Data Analysis and Graphics Using R", plus other
	selected datasets and functions.
}
\details{
For a list of , use \code{library(help="DAAG")}.
}
\author{
Author: John H Maindonald

Maintainer: W John Braun \email{braun@stats.uwo.ca}
}
\keyword{package}
