\name{frostedflakes}
\alias{frostedflakes}
\docType{data}
\title{Frosted Flakes data}
\usage{elastic1}
\description{
The \code{frosted flakes} data frame has 101 rows and 2 columns
giving the sugar concentration (in percent) for 25 g samples
of a cereal as measured by 2 methods -- high performance liquid
chromatography (a slow accurate lab method) and a quick method
using the infra-analyzer 400.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Lab}{careful laboratory analysis measurements using
high performance liquid chromatography}
    \item{IA400}{measurements based on the infra-analyzer 400}
  }
}
\source{
 W. J. Braun
}
\keyword{datasets}
