\name{measles}
\alias{measles}
\docType{data}
\title{Deaths in London from measles}
\description{
  Deaths in London from measles: 1629 -- 1939, with gaps.
}
\usage{data(measles)}
\format{
  The format is:
 Time-Series [1:311] from 1629 to 1939: 42 2 3 80 21 33 27 12 NA NA ...
}
\source{
  Guy, W. A. 1882. Two hundred and fifty years of small pox in London.
  Journal of the Royal Statistical Society 399-443.

  Stocks, P. 1942. Measles and whooping cough during the dispersal of 
1939-1940. Journal of the Royal Statistical Society 105:259-291.
}
\references{
  Lancaster, H. O. 1990. Expectations of Life. Springer.
}
\keyword{datasets}
