\name{ozone}
\alias{ozone}
\non_function{}
\title{Ozone Data}
\usage{data(ozone)}
\description{
The \code{ozone} data frame has 45 rows and 11 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Year}{a numeric vector}
    \item{Aug}{a numeric vector}
    \item{Sep}{a numeric vector}
    \item{Oct}{a numeric vector}
    \item{Nov}{a numeric vector}
    \item{Dec}{a numeric vector}
    \item{Jan}{a numeric vector}
    \item{Feb}{a numeric vector}
    \item{Mar}{a numeric vector}
    \item{Apr}{a numeric vector}
    \item{Year}{a numeric vector}
  }
}
\source{

}
\references{

}
\examples{
data(ozone)
}
\keyword{datasets}
