\name{rareplants}
\alias{rareplants}
\title{Rare and Endangered Plant Species}
\usage{rareplants}
\description{
These data were taken from species lists for South Australia, Victoria and 
Tasmania.  Species were classified as CC, CR, RC and RR, with C denoting
common and R denoting rare.  The first code relates to South Australia 
and Victoria, and the second to Tasmania.  They were further classified
by habitat according to the Victorian register, where D = dry only, 
W = wet only, and WD = wet or dry.
}
\format{
  The format is:
 chr "rareplants"
}
\source{
Jasmyn Lynch, Department of Botany and Zoology at Australian National 
University
}
\examples{
chisq.test(rareplants)
}
\keyword{datasets}
\keyword{datasets}
