\name{SP500close}
\alias{SP500close}
\docType{data}
\title{Closing Numbers for S and P 500 Index}
\description{
  Closing numbers for S and P 500 Index, Jan. 1, 1990 through
early 2000.  
}
\usage{SP500close}
\source{
  Derived from SP500 in the MASS library.
}
\examples{
ts.plot(SP500close)
}
\keyword{datasets}
