\name{SP500W90}
\alias{SP500W90}
\docType{data}
\title{Closing Numbers for S and P 500 Index - First 100 Days of 1990}
\description{
  Closing numbers for S and P 500 Index, Jan. 1, 1990 through
early 2000.  
}
\usage{SP500W90}
\source{
  Derived from SP500 in the MASS library.
}
\examples{
ts.plot(SP500W90)
}
\keyword{datasets}
