\name{stVincent}
\alias{stVincent}
\title{Averages by block of yields for the St. Vincent Corn data}
\description{
  These data frames have yield averages by blocks (parcels).
}
\usage{stVincent}
\format{ A data frame with 324 observations on 8 variables.
  \describe{
    \item{code}{a numeric vector}
    \item{island}{a numeric vector}
    \item{id}{a numeric vector}
    \item{site}{a factor with 8 levels.}
    \item{block}{a factor with levels \code{I} \code{II} \code{III} \code{IV}}
    \item{plot}{a numeric vector}
    \item{trt}{a factor consisting of 12 levels}
    \item{harvwt}{a numeric vector; the average yield}
  }
}
\source{
  Andrews DF; Herzberg AM, 1985. Data. A Collection of Problems from
  Many Fields for the Student and Research Worker. Springer-Verlag.
  (pp. 339-353)
}
\keyword{datasets}
