\name{vince111b}
\alias{vince111b}
\title{Averages by block of corn yields, for treatment 111 only}
\description{
  These data frames have averages by blocks (parcels) for the treatment
  \code{111}.
}
\usage{vince111b}
\format{ A data frame with 36 observations on 8 variables.
  \describe{
    \item{site}{a factor with levels 
      \code{AGSV} \code{CASV} \code{CPSV}
      \code{LPSV} \code{MPSV} \code{OOSV} \code{OTSV} \code{SSSV} \code{UISV}
    }
    \item{parcel}{a factor with levels \code{I} \code{II} \code{III} \code{IV}}
    \item{code}{a numeric vector}
    \item{island}{a numeric vector}
    \item{id}{a numeric vector}
    \item{plot}{a numeric vector}
    \item{trt}{a numeric vector}
    \item{harvwt}{a numeric vector}
  }
}
\source{
  Andrews DF; Herzberg AM, 1985. Data. A Collection of Problems from
  Many Fields for the Student and Research Worker. Springer-Verlag.
  (pp. 339-353)
}
\keyword{datasets}
