\name{DAAG-package}
\alias{DAAG-package}
\alias{DAAG}
\docType{package}
\title{
The R DAAG Package
}
\description{
Various data sets and functions used or referred to in the book
        Maindonald, J.H. and Braun, W.J. (2nd edn 2007)  
        "Data Analysis and Graphics Using R", plus other
	selected datasets and functions.
}
\details{
For a list of , use \code{library(help="DAAG")}.
}
\author{
Author: John Maindonald
  
Maintainer: John Maindonald  \email{john.maindonald@anu.edu.au}
}
\keyword{package}
