\name{fossilfuel}
\alias{fossilfuel}
\title{Fossil Fuel Data}
\usage{fossilfuel}
\description{
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{year}{a numeric vector giving the year the measurement
was taken.}
    \item{carbon}{a numeric vector giving the total worldwide
carbon emissions from fossil fuel use, in millions of tonnes.}
  }
}
\source{
Marland et al (2003)
}
\examples{
plot(fossilfuel)
}
\keyword{datasets}
