\name{frostedflakes}
\alias{frostedflakes}
\non_function{}
\title{Frosted Flakes data}
\usage{elastic1}
\description{
The \code{frosted flakes} data frame has 101 rows and 2 columns
giving the sugar concentration (in unspecified units) for samples
of a cereal as measured by 2 methods.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Lab}{careful laboratory analysis measurements}
    \item{HPLC}{high performance liquid chromatography measurements}
  }
}
\source{
 W. J. Braun
}
\keyword{datasets}
