\name{hills2000}
\alias{hills2000}
\non_function{}
\title{Scottish Hill Races Data - 2000}
\usage{data(hills2000)}
\description{
The record times in 2000 for 77 Scottish hill races.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{h}{a numeric vector}
    \item{m}{a numeric vector}
    \item{s}{a numeric vector}
    \item{h0}{a numeric vector}
    \item{m0}{a numeric vector}
    \item{s0}{a numeric vector}
    \item{dist}{distance, in miles (on the map)}
    \item{climb}{total height gained during the route, in feet}
    \item{time}{record time in hours}
    \item{timef}{record time in hours for females}
    \item{type}{a factor, with levels indicating type of race}
  }
}
\source{
?
}
\examples{
data(hills2000)
}
\keyword{datasets}
