\name{leaftemp.all}
\alias{leaftemp.all}
\title{Full Leaf and Air Temperature Data Set}
\usage{leaftemp.all}
\description{
The \code{leaftemp.all} data frame has 62 rows and 9 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{glasshouse}{a factor with levels
\code{A},
\code{B},
\code{C} }
    \item{CO2level}{a factor with Carbon Dioxide Levels:
\code{high},
\code{low},
\code{medium} }
    \item{day}{a factor}
    \item{light}{a numeric vector}
    \item{CO2}{a numeric vector}
    \item{tempDiff}{Difference between Leaf and Air Temperature}
    \item{BtempDiff}{a numeric vector}
    \item{airTemp}{Air Temperature}
    \item{vapPress}{Vapour Pressure}
  }
}
\source{
J.H. Maindonald
}
\keyword{datasets}
